+++ src/AtomUI.Desktop.Controls/Space/Space.SemanticParts.cs
@@ -1,0 +1,24 @@
+using AtomUI.Theme;
+using Avalonia.Controls;
+
+namespace AtomUI.Desktop.Controls;
+
+[SemanticPart(
+    "item",
+    SelectorClass = "semantic-item",
+    SelectorRoute = "> .semantic-item",
+    ContractType = typeof(Control),
+    Cardinality = SemanticPartCardinality.Multiple,
+    Since = "6.0",
+    RuntimeCreated = true)]
+[SemanticPart(
+    "separator",
+    SelectorClass = "semantic-separator",
+    SelectorRoute = "> .semantic-separator",
+    ContractType = typeof(Control),
+    Cardinality = SemanticPartCardinality.Multiple,
+    Since = "6.0",
+    RuntimeCreated = true)]
+public partial class Space
+{
+}